hendricks-request-file-bas
This is used specifically in the escrow tax department..  to match up our current listing of parcels created via a report.   I made this to significantly expedite our antiquated process of eyeball comparing our current listing of 1,000's of parcels to an old listing the county sends us.  There may be 100's of  new parcels and 100's of deleted ones.  This allows us to instantly see which are new and add them to the county listing.  It has some stringent requirements on columns, so is probably useless to an outside user, but may prove useful for anyone looking for inspiration on how to go about it.

This is used specifically on a format used very heavily in Indiana, they use an Excel spreadsheet for the request files and payment files.  My bank services quite a few properties in Indiana, so it has saved me a lot of time.