Create Hendricks-Payment-File.bas
This is run for the payment file that is sent by the county, typically a month or so after this is generated.  I use a newer database model, and it first does an add/delete filter, for any new parcels or paid off loans that may have occurred between our "request file" and our "payment file".  

It is up to us to manually add in the "name of the property owner, and address, and track down the amounts from the county that are due, if we do indeed we wish to pay via automation and include it on this spreadsheet.